requirements.txt
pandas
gensim
nltk
python-twitter
spacy
gsheets
pyLDAvis
matplotlib
scikit-learn
networkx
google-auth-httplib2
google-api-python-client
sklearn
xmltodict
sendgrid
BeautifulSoup4
html5lib
tensorflow
yapf
coverage
codecov
